Temp

+++ src/Vanke.WX.Weixin/Controllers/DinnerPlacesController.cs
@@ -28,7 +28,7 @@ namespace Vanke.WX.Weixin.Controllers
         /// <param name="id"></param>
         /// <returns></returns>
         [HttpGet]
-        public async Task<DinnerPlace> Get(int id)
+        public async Task<DinnerPlace> Get(long id)
         {
             return await _dinnerPlaceService.GetByKeyAsync(id);
         }
@@ -57,7 +57,7 @@ namespace Vanke.WX.Weixin.Controllers
         /// <param name="id"></param>
         /// <returns></returns>
         [HttpDelete]
-        public async Task Delete(int id)
+        public async Task Delete(long id)
         {
             await _dinnerPlaceService.RemoveAsync(id);
         }
